Extracted RecipientLookUpManager to a separate class. Refactored code.| #74
